Enhance subgame solving and fallback strategy in API
- Implemented a fallback warm start strategy in the API for subgame solving when blueprint lookup fails, improving robustness in decision-making.
- Added detailed logging for fallback strategy generation and subgame edge handling to aid in debugging.
- Updated the `Decision` struct to handle edge parsing more effectively.
- Refactored the `solve_subgame_unsafe` method to accommodate the new fallback strategy logic.
- Adjusted the handling of warm start strategies in the `SubgameProfile` to ensure proper initialization from available data.

+++ src/analysis/response.rs
@@ -75,7 +75,7 @@ impl From<tokio_postgres::Row> for Sample {
 impl From<tokio_postgres::Row> for Decision {
     fn from(row: tokio_postgres::Row) -> Self {
         Self {
-            edge: Edge::from(row.get::<_, i64>("edge") as u64).to_string(),
+            edge: Edge::from(row.get::<_, i64>("edge") as u8).to_string(),
             prob: Probability::from(row.get::<_, f32>("policy")),
         }
     }
